Refactor config file name and setup process
